+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## nbfc-linux-0.1.1 (2012-05-02)
+- Fixed type `--config-fie` to `--config-file`
+- Using default temperature thresholds if those in config are empty or not present.  [Issue #2](https://github.com/braph/nbfc-linux/issues/2#issue-897727519)
+
 ## nbfc-linux-0.1.0 (2021-05-19)
 - Added CHANGELOG.md
 - Added --version flags to all programs